# Conflicts:
#	src/main/java/com/team2502/robot2017/DashboardData.java
#	src/main/java/com/team2502/robot2017/Robot.java
#
src/main/java/com/team2502/robot2017/command/autonomous/AutoCommandG2.ja
va
#
src/main/java/com/team2502/robot2017/command/autonomous/AutoVCommand.jav
a
#
src/main/java/com/team2502/robot2017/subsystem/DistanceSensorSubsystem.j
ava

+++ src/main/java/com/team2502/robot2017/command/DriveTillDistanceCommand.java
@@ -31,7 +31,7 @@ public class DriveTillDistanceCommand extends Command
 			driveTrain.leftSpeed = -0.5;
 			driveTrain.rightSpeed = 0.5;
 			
-			Distance = Sensor.getSensorDistance();
+			Distance = (double) Sensor.getSensorDistance();
 			if (Distance > 14)
 			{
 				driveTrain.runMotors(driveTrain.leftSpeed, driveTrain.rightSpeed);
